Change: abort if blib is already installed

+++ Makefile
@@ -16,6 +16,7 @@ test : $(TESTS)
 	done
 
 install : blib.oo.sh deps/libblib tests
+	[ -L "$(BINPATH)" -o -f "$(BINPATH)" ] && { echo "Already installed. Aborting" >&2; exit;}; \
 	[ -d "$(ROOT)" ] || { echo "Making directory..."; mkdir "$(ROOT)"; }; \
 	cp -r $^ $(ROOT)/ ;\
 	{ cd $(ROOT); \
